+++ packages/expo-dev-launcher/bundle/index.ts
@@ -3,6 +3,14 @@ import { enableScreens } from 'react-native-screens';
 
 import { App } from './App';
 
+// the bundle produced by RN 77 doesn't have set and clear -Immediate when running on 76
+if (!global.setImmediate) {
+  global.setImmediate = require('./immediateShim').setImmediate;
+}
+if (!global.clearImmediate) {
+  global.clearImmediate = require('./immediateShim').clearImmediate;
+}
+
 // Hide this target from the JS inspector
 globalThis.__expo_hide_from_inspector__ = 'expo-dev-launcher';