+++ packages/expo-updates/ios/EXUpdates/Procedures/CheckForUpdateProcedure.swift
@@ -40,7 +40,7 @@ final class CheckForUpdateProcedure: StateMachineProcedure {
   }
 
   func run(procedureContext: ProcedureContext) {
-    procedureContext.processStateEvent(UpdatesStateEventCheck())
+    procedureContext.processStateEvent(.check)
 
     database.databaseQueue.async {
       let embeddedUpdate = EmbeddedAppLoader.embeddedManifest(withConfig: self.config, database: self.database)
@@ -63,20 +63,20 @@ final class CheckForUpdateProcedure: StateMachineProcedure {
             switch updateDirective {
             case is NoUpdateAvailableUpdateDirective:
               self.successBlock(CheckForUpdateResult.noUpdateAvailable(reason: RemoteCheckResultNotAvailableReason.noUpdateAvailableOnServer))
-              procedureContext.processStateEvent(UpdatesStateEventCheckComplete())
+              procedureContext.processStateEvent(.checkCompleteUnavailable)
               procedureContext.onComplete()
               return
             case let rollBackUpdateDirective as RollBackToEmbeddedUpdateDirective:
               if !self.config.hasEmbeddedUpdate {
                 self.successBlock(CheckForUpdateResult.noUpdateAvailable(reason: RemoteCheckResultNotAvailableReason.rollbackNoEmbedded))
-                procedureContext.processStateEvent(UpdatesStateEventCheckComplete())
+                procedureContext.processStateEvent(.checkCompleteUnavailable)
                 procedureContext.onComplete()
                 return
               }
 
               guard let embeddedUpdate = embeddedUpdate else {
                 self.successBlock(CheckForUpdateResult.noUpdateAvailable(reason: RemoteCheckResultNotAvailableReason.rollbackNoEmbedded))
-                procedureContext.processStateEvent(UpdatesStateEventCheckComplete())
+                procedureContext.processStateEvent(.checkCompleteUnavailable)
                 procedureContext.onComplete()
                 return
               }
@@ -88,15 +88,13 @@ final class CheckForUpdateProcedure: StateMachineProcedure {
                 filters: manifestFilters
               ) {
                 self.successBlock(CheckForUpdateResult.noUpdateAvailable(reason: RemoteCheckResultNotAvailableReason.rollbackRejectedBySelectionPolicy))
-                procedureContext.processStateEvent(UpdatesStateEventCheckComplete())
+                procedureContext.processStateEvent(.checkCompleteUnavailable)
                 procedureContext.onComplete()
                 return
               }
 
               self.successBlock(CheckForUpdateResult.rollBackToEmbedded(commitTime: rollBackUpdateDirective.commitTime))
-              procedureContext.processStateEvent(
-                UpdatesStateEventCheckCompleteWithRollback(rollbackCommitTime: rollBackUpdateDirective.commitTime)
-              )
+              procedureContext.processStateEvent(.checkCompleteWithRollback(rollbackCommitTime: rollBackUpdateDirective.commitTime))
               procedureContext.onComplete()
               return
             default:
@@ -106,7 +104,7 @@ final class CheckForUpdateProcedure: StateMachineProcedure {
 
           guard let update = updateResponse.manifestUpdateResponsePart?.updateManifest else {
             self.successBlock(CheckForUpdateResult.noUpdateAvailable(reason: RemoteCheckResultNotAvailableReason.noUpdateAvailableOnServer))
-            procedureContext.processStateEvent(UpdatesStateEventCheckComplete())
+            procedureContext.processStateEvent(.checkCompleteUnavailable)
             procedureContext.onComplete()
             return
           }
@@ -138,7 +136,7 @@ final class CheckForUpdateProcedure: StateMachineProcedure {
 
           if shouldLaunch {
             self.successBlock(CheckForUpdateResult.updateAvailable(manifest: update.manifest.rawManifestJSON()))
-            procedureContext.processStateEvent(UpdatesStateEventCheckCompleteWithUpdate(manifest: update.manifest.rawManifestJSON()))
+            procedureContext.processStateEvent(.checkCompleteWithUpdate(manifest: update.manifest.rawManifestJSON()))
             procedureContext.onComplete()
             return
           }
@@ -147,11 +145,11 @@ final class CheckForUpdateProcedure: StateMachineProcedure {
             RemoteCheckResultNotAvailableReason.updatePreviouslyFailed :
             RemoteCheckResultNotAvailableReason.updateRejectedBySelectionPolicy
           self.successBlock(CheckForUpdateResult.noUpdateAvailable(reason: reason))
-          procedureContext.processStateEvent(UpdatesStateEventCheckComplete())
+          procedureContext.processStateEvent(.checkCompleteUnavailable)
           procedureContext.onComplete()
           return
         } errorBlock: { error in
-          procedureContext.processStateEvent(UpdatesStateEventCheckError(message: error.localizedDescription))
+          procedureContext.processStateEvent(.checkError(errorMessage: error.localizedDescription))
           self.successBlock(CheckForUpdateResult.error(error: error))
           procedureContext.onComplete()
           return